# perf(python): further optimize datetime conversion